requirements.txt
requests
feedparser
python-telegram-bot>=20.4
pyyaml